+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## Unreleased
+
+### Bug Fixes
+
+- [\#324](https://github.com/cosmos/iavl/pull/324) Fix `DeleteVersions` causing a memory leak. 
+  Add `DeleteVersionsRange` to delete in range (@klim0v)
+
 ## 0.14.2 (October 12, 2020)
 
 ### Bug Fixes